refactor(vscode-extension): split extension.ts into service modules with a shared context seam
